New build system working. See comments in build.xml for instructions.
The app needs comprehensive regression testing, especially to make sure all the suporting files are
properly accessible.

The crypto lib files jce.jar and jsse.jar were apparently moved within the JDK for Java 8. To avoid a
dependency on a specific Java 8 update version (and because the crypto jars are only avaiable in Oracle
Java (e.g. not OpenJDK), the jars were added to the app's folder. We can re-evaluate if they need to remain
there when we upgrade to Java 10 or 11.

The build system does not support a self-contained bundle without an embedded JRE. The choice is to either
build a launchable jar file or an app bundle with an embedded JRE. We have the latter.

The app bundle needs to be signed, to avoid running afoul of gatekeeper.

The readme documentation will be updated when the release is compeltely ready.